requirements.txt
python-dotenv
weasyprint
